Add include_dirs to servedocs
This patch add a new keyword argument `include_dirs` to `servedocs`
which make it possible to add additional source directories (in addition
to `docs/src`) that should be considered for the file watching. This
does, for example, make it possible to watch the source files of your
package, and call `Revise.revise()` in `docs/make.jl` in order to
automatically update docstrings.