Allow letting the docs build to fail with servedocs() (#176)
Previously `servedocs()` would exit whenever the docs failed to build, and it
can be annoying to always restart it on errors. Now when building the docs fails
a special status will be set on the `SimpleWatcher` which will get
`serve_file()` to return a custom error page (which will be reloaded
automatically when the docs build again).