Remove Pkg as a dependency
In recent Julia versions Pkg is not in the sysimage and therefore it can
not be loaded "for free". Pkg was only used to conditionally activate
the documentation environment in `LiveServer.servedocs` so the loss of
functionality is basically non-existent -- it is easy enough to
configure the package environment before calling `LiveServer.servedocs`.

Timings for loading LiveServer on Julia 1.12:
```
julia> @time using LiveServer # this PR
  0.448249 seconds

julia> @time using LiveServer # master
  0.743163 seconds
```